DEV  Moved the dialog that showd Network info to class NetworkDialog
